Put Virgil into its own profile
If Virgil is in the default profile, it watches Java code when we want
to just run the tests. Its asynchronous code (re)loading logic leads to
a race condition -ish situation where we're trying to load a namespace
that has not been completely reloaded. The simplest thing for making
tests work is to put it into its own profile.