Add tagged value module
This aims to open the doors for jsonista to become a viable alternative
for people currently using EDN with transit by providing a way to encode
EDN types using a tagged JSON list. Currently, only keywords are
supported.

Regarding `TaggedValueOrPersistentVectorDeserializer`, it's worth noting
that it doesn't just peek for a tag and then fall back on the default
`PersistentVectorDeserializer`, since peeking seemed to require a call
to `nextValue()` and I found no way to backtack. So it seems I need to
try building the transient and then just special-case the zeroth element
to check for the tag.

Right now, it doesn't do anything special with mapping tags to different
types in the Java code, since there's only one and I figure that change
can be made when it's needed. The Clojure API, however, remains more
general, since the tag needs to come in as a key.  This will make adding
new keys to the API trivial, if more types are supported in the future.

This closes #35.