Add META-INF with native-image.properties
As `clj-easy/graal-build-time` states:

> From GraalVM 22 onwards, the native-image global --initialize-at-build-time option will be deprecated. This means you will have to list every package that you want to initialize at build time separately, like initialize-at-build-time=clojure,my_library

Without now discouraged `--initialize-at-build-time` with no arguments passed the `native-image` tool fails to build the project that depends on `metosin/jsonista`. This is due to the fact, that some of the `com.fasterxml.jackson` classes are initialized unconditionally in `core.clj`.

The following changes auto provides the necessary `jsonista` configurations to GraalVM `native-image`, are backward compatible for all GraalVM `jsonista` users, and should not introduce any breakage.